Grab bag of minor fixes and changes from Adobe's Flash Player team:
-- MIR: improve int/uint compares in AMD64 mode by using a 64-bit register directly
-- Vector: add missing DWB macro to m_array member
-- various fixes for compilation on ARM CPU
-- fixes for various compiler warnings (e.g., "uint >=0")
-- renamed the VS2008 project/solution files from "_9" to "2008", as this was the nomenclature the Flash team is using, and AFAIK no one else is using the VS2008 projects at this time. Also tweaked some settings in the VS2008 files themselves.

Note that diff has lots of "changes" that are non-changes, apparently due to line-ending weirdness. Anyone know how to get hg to ignore line-ending changes when doing diffs?

+++ core/AvmCore.h
@@ -230,7 +230,7 @@ const int kBufferPadding = 16;
 		 * will check the stack pointer to make sure it
 		 * doesn't go below this value.
 		 */
-		uint32 minstack;
+		uintptr_t minstack;
 
 		/**
 		 * This method will be invoked when the first exception
